BugFix&Piece

+++ JeuEchecs/src/com/echecs/pieces/Dame.java
@@ -11,7 +11,7 @@ public class Dame extends Piece
     public boolean peutSeDeplacer(Position pos1, Position pos2, Piece[][] echiquier)
     {
         //Deplacement en X et Y
-        int depX = pos2.getColonne() - pos1.getColonne();
+        int depX = (int)pos2.getColonne() - (int)pos1.getColonne();
         int depY = pos2.getLigne() - pos1.getLigne();
         int incrementX = depX > 0 ? 1 : -1;
         int incrementY = depY > 0 ? 1 : -1;
